requirements.txt
beautifulsoup4
python-dotenv
py-cord